vns com problema

+++ header/Grafo.hpp
@@ -29,7 +29,7 @@ public:
   int quantidadeNos();
   bool noHotel(int id);
   int tamanhoTrip();
-  int calculaCustoTempo(std::vector<int> solucao);
+  float calculaCustoTempo(std::vector<int> solucao);
   std::unordered_map<int, Vertice> copiarNos() {
     return nos;
   }
